+++ tests/unit/s2n_handshake_test.c
@@ -81,7 +81,9 @@ static int try_handshake(struct s2n_connection *server_conn, struct s2n_connecti
     return 0;
 }
 
-int test_cipher_preferences(struct s2n_config *server_config, struct s2n_config *client_config) {
+int test_cipher_preferences(struct s2n_config *server_config, struct s2n_config *client_config,
+    uint8_t client_protocol_version)
+{
     const struct s2n_cipher_preferences *cipher_preferences;
 
     cipher_preferences = server_config->cipher_preferences;
@@ -135,7 +137,7 @@ int test_cipher_preferences(struct s2n_config *server_config, struct s2n_config 
         GUARD(s2n_connection_set_write_fd(client_conn, client_to_server[1]));
         GUARD(s2n_connection_set_config(client_conn, client_config));
         client_conn->server_protocol_version = S2N_TLS12;
-        client_conn->client_protocol_version = S2N_TLS12;
+        client_conn->client_protocol_version = client_protocol_version;
         client_conn->actual_protocol_version = S2N_TLS12;
 
         GUARD(s2n_connection_set_read_fd(server_conn, client_to_server[0]));
@@ -227,7 +229,7 @@ int main(int argc, char **argv)
         
         EXPECT_SUCCESS(s2n_config_set_verification_ca_location(client_config, S2N_DEFAULT_TEST_CERT_CHAIN, NULL));
 
-        EXPECT_SUCCESS(test_cipher_preferences(server_config, client_config));
+        EXPECT_SUCCESS(test_cipher_preferences(server_config, client_config, S2N_TLS12));
 
         EXPECT_SUCCESS(s2n_cert_chain_and_key_free(chain_and_key));
         EXPECT_SUCCESS(s2n_config_free(server_config));
@@ -265,7 +267,45 @@ int main(int argc, char **argv)
 
         EXPECT_SUCCESS(s2n_config_set_verification_ca_location(client_config, S2N_ECDSA_P384_PKCS1_CERT_CHAIN, NULL));
         
-        EXPECT_SUCCESS(test_cipher_preferences(server_config, client_config));
+        EXPECT_SUCCESS(test_cipher_preferences(server_config, client_config, S2N_TLS12));
+
+        EXPECT_SUCCESS(s2n_cert_chain_and_key_free(chain_and_key));
+        EXPECT_SUCCESS(s2n_config_free(server_config));
+        free(cert_chain_pem);
+        free(private_key_pem);
+        free(dhparams_pem);
+
+    }
+
+    /*  test_with_rsa_cert() with client "supported_version" as TLS1.3 and server protocol version TLS1.2; */
+    {
+        struct s2n_config *server_config, *client_config;
+        char *cert_chain_pem;
+        char *private_key_pem;
+        char *dhparams_pem;
+        struct s2n_cert_chain_and_key *chain_and_key;
+
+        EXPECT_NOT_NULL(cert_chain_pem = malloc(S2N_MAX_TEST_PEM_SIZE));
+        EXPECT_NOT_NULL(private_key_pem = malloc(S2N_MAX_TEST_PEM_SIZE));
+        EXPECT_NOT_NULL(dhparams_pem = malloc(S2N_MAX_TEST_PEM_SIZE));
+
+        EXPECT_NOT_NULL(server_config = s2n_config_new());
+
+        EXPECT_SUCCESS(s2n_read_test_pem(S2N_DEFAULT_TEST_CERT_CHAIN, cert_chain_pem, S2N_MAX_TEST_PEM_SIZE));
+        EXPECT_SUCCESS(s2n_read_test_pem(S2N_DEFAULT_TEST_PRIVATE_KEY, private_key_pem, S2N_MAX_TEST_PEM_SIZE));
+        EXPECT_SUCCESS(s2n_read_test_pem(S2N_DEFAULT_TEST_DHPARAMS, dhparams_pem, S2N_MAX_TEST_PEM_SIZE));
+        EXPECT_NOT_NULL(chain_and_key = s2n_cert_chain_and_key_new());
+        EXPECT_SUCCESS(s2n_cert_chain_and_key_load_pem(chain_and_key, cert_chain_pem, private_key_pem));
+        EXPECT_SUCCESS(s2n_config_add_cert_chain_and_key_to_store(server_config, chain_and_key));
+        EXPECT_SUCCESS(s2n_config_add_dhparams(server_config, dhparams_pem));
+
+        EXPECT_SUCCESS(s2n_config_set_cipher_preferences(server_config, "test_some_rsa"));
+
+        client_config = s2n_fetch_unsafe_client_testing_config();
+
+        EXPECT_SUCCESS(s2n_config_set_verification_ca_location(client_config, S2N_DEFAULT_TEST_CERT_CHAIN, NULL));
+
+        EXPECT_SUCCESS(test_cipher_preferences(server_config, client_config, S2N_TLS13));
 
         EXPECT_SUCCESS(s2n_cert_chain_and_key_free(chain_and_key));
         EXPECT_SUCCESS(s2n_config_free(server_config));
