+++ tls/s2n_client_key_exchange.c
@@ -88,9 +88,7 @@ static int calculate_keys(struct s2n_connection *conn, struct s2n_blob *shared_k
 
 int s2n_rsa_client_key_recv(struct s2n_connection *conn, struct s2n_blob *shared_key)
 {
-    // printf("OMG s2n_rsa_client_key_recv client hello version %d\n", s2n_connection_get_client_hello_version(conn));
     struct s2n_stuffer *in = &conn->handshake.io;
-    // uint8_t client_protocol_version[S2N_TLS_PROTOCOL_VERSION_LEN];
     uint8_t client_hello_version[S2N_TLS_PROTOCOL_VERSION_LEN];
     uint16_t length;
 
@@ -102,9 +100,10 @@ int s2n_rsa_client_key_recv(struct s2n_connection *conn, struct s2n_blob *shared
 
     S2N_ERROR_IF(length > s2n_stuffer_data_available(in), S2N_ERR_BAD_MESSAGE);
 
-    /* Keep a copy of the client protocol version in wire format */
-    // client_protocol_version[0] = conn->client_protocol_version / 10;
-    // client_protocol_version[1] = conn->client_protocol_version % 10;
+    /* Keep a copy of the client hello version in wire format, which should be
+     * either the protocol versin supported by client if the supported version is <= TLS1.2,
+     * or TLS1.2 (the legacy version) if the supported version is TLS1.3
+     */
     client_hello_version[0] = conn->client_hello_version / 10;
     client_hello_version[1] = conn->client_hello_version % 10;
 
@@ -124,12 +123,8 @@ int s2n_rsa_client_key_recv(struct s2n_connection *conn, struct s2n_blob *shared
     /* Set rsa_failed to 1 if s2n_pkey_decrypt returns anything other than zero */
     conn->handshake.rsa_failed = !!s2n_pkey_decrypt(conn->handshake_params.our_chain_and_key->private_key, &encrypted, shared_key);
 
-    // printf("OMG s2n_rsa_client_key_recv rsa_failed 1st is %d\n", conn->handshake.rsa_failed);
-
     /* Set rsa_failed to 1, if it isn't already, if the protocol version isn't what we expect */
     conn->handshake.rsa_failed |= !s2n_constant_time_equals(client_hello_version, shared_key->data, S2N_TLS_PROTOCOL_VERSION_LEN);
-
-    // printf("OMG s2n_rsa_client_key_recv rsa_failed 2nd is %d\n", conn->handshake.rsa_failed);
 
     return 0;
 }
@@ -212,12 +207,6 @@ int s2n_ecdhe_client_key_send(struct s2n_connection *conn, struct s2n_blob *shar
 
 int s2n_rsa_client_key_send(struct s2n_connection *conn, struct s2n_blob *shared_key)
 {
-    // uint8_t client_protocol_version[S2N_TLS_PROTOCOL_VERSION_LEN];
-    // client_protocol_version[0] = conn->client_protocol_version / 10;
-    // client_protocol_version[1] = conn->client_protocol_version % 10;
-
-    // printf("OMG s2n_rsa_client_key_send client hello version %d\n", s2n_connection_get_client_hello_version(conn));
-    // printf("OMG s2n_rsa_client_key_send actual protocol version %d\n", s2n_connection_get_actual_protocol_version(conn));
     uint8_t client_hello_version[S2N_TLS_PROTOCOL_VERSION_LEN];
     client_hello_version[0] = conn->client_hello_version / 10;
     client_hello_version[1] = conn->client_hello_version % 10;
@@ -227,7 +216,10 @@ int s2n_rsa_client_key_send(struct s2n_connection *conn, struct s2n_blob *shared
 
     GUARD(s2n_get_private_random_data(shared_key));
 
-    /* Over-write the first two bytes with the client protocol version, per RFC2246 7.4.7.1 */
+    /* Over-write the first two bytes with the client hello version, per RFC2246 7.4.7.1.
+     * this client hello version is either the "supported_versions" extension by client if
+     * that supported version is <= TLS1.2, or TLS1.2 (legacy version) if the supported version is TLS1.3
+     */
     memcpy_check(conn->secure.rsa_premaster_secret, client_hello_version, S2N_TLS_PROTOCOL_VERSION_LEN);
 
     int encrypted_size = s2n_pkey_size(&conn->secure.server_public_key);
